sched: cpudl remove CPUDL_INVALID and a fix.
CPUDL_INVALID is meaningless using u64 deadlines. This commit
removes it and insert a flag in cpudl_set to be able to
disciminate valid deadlines (when the last -dl task in a rq
finishes we set the flag to 0 to indicate the node has to be
removed from the cpudl heap).
The fix addresses the case when, after a node removal, deadline of
the repacing last element is higher than its new parent. In this
case cpudl_heapify didn't work.