Make dl_bw a sub-quota of rt_bw
Change real-time bandwidth management as to make dl_bw a sub-quota
of rt_bw. This commit leaves rt_bw at its default value and sets
dl_bw at 40% of the former.
Checks are made when the user tries to change dl_bw sub-quota as to
not fall below what currently used.
Similar checks are to be made when the user modifies rt_bw (dl_bw
actual quota is a function of it) and this is a TODO.